Abstract our renaming script.

+++ update_theme_name.sh
@@ -2,10 +2,10 @@
 
 # our own little yeoman
 #
-# This script works on a theme that was started by cloning ts_motif.
-# If the theme still has the name ts_motif sprinkled throughout, this does the
-# job of changing those instances of 'ts_motif' to the new theme name.
-# It also renames the directory from ts_motif to the new theme name if it is
+# This script works on a theme that was started by cloning renamed_theme.
+# If the theme still has the name renamed_theme sprinkled throughout, this does the
+# job of changing those instances of 'renamed_theme' to the new theme name.
+# It also renames the directory from renamed_theme to the new theme name if it is
 # not already re-named.
 #
 # call like this:
@@ -20,58 +20,62 @@
   fi
 
   echo Theme name is $THEME_NAME
-  # assumes the currently running script is inside the ts_motif directory that will be re-named
+  # assumes the currently running script is inside the renamed_theme directory that will be re-named
   THEME_DEST="$( cd "$(dirname "$0")" ; cd ..; pwd -P )"/$THEME_NAME
   CURRENT_DIR="$(pwd -P )"
   echo "Theme will be placed in $THEME_DEST"
 
   if [ $THEME_DEST != $CURRENT_DIR ]
   then
+    echo Renaming the theme directory
     mv $CURRENT_DIR $THEME_DEST
   fi
 
-  PROJECT_NAME=$(echo $THEME_DEST | sed 's:.*/\(.*\)/web/.*:\1:g')
+  # Go through files and edit them replacing renamed_theme with the new theme name
+  # The "sed" command replaced the thing between "s/" and the middle "/" with the thing
+  # after the middle "/" and the "/g". Its used both to generate new filenames and to
+  # edit the contents of files.
 
-  # go through files and edit them replacing ts_motif with the new theme name
-  sed "s/ts_motif_/${THEME_NAME}_/g" $THEME_DEST/ts_motif.theme >$THEME_DEST/$THEME_NAME.theme
-  rm $THEME_DEST/ts_motif.theme
+  for file in $THEME_DEST/*renamed_theme*
+    do
+      echo Renaming ${file}
+      newfile="$(echo ${file} | sed -e "s/renamed_theme/${THEME_NAME}/g")" ;
+      mv $file $newfile
+  done
 
-  sed "s/ts_motif./${THEME_NAME}./g" $THEME_DEST/ts_motif.breakpoints.yml >$THEME_DEST/$THEME_NAME.info.yml
-  rm $THEME_DEST/ts_motif.breakpoints.yml
-
-  sed "s/name: Motif/name: ${THEME_NAME}/g" $THEME_DEST/ts_motif.info.yml | sed "s/ts_motif/${THEME_NAME}/g" >$THEME_DEST/$THEME_NAME.info.yml
-  rm $THEME_DEST/ts_motif.info.yml
-
-  sed "s/ts_motif/${THEME_NAME}/g" $THEME_DEST/ts_motif.libraries.yml >$THEME_DEST/$THEME_NAME.libraries.yml
-  rm $THEME_DEST/ts_motif.libraries.yml
-
-  mv $THEME_DEST/composer.json $THEME_DEST/composer.child
-  sed "s/ts_motif/${THEME_NAME}/g" $THEME_DEST/composer.child >$THEME_DEST/composer.json
-  rm $THEME_DEST/composer.child
-
-  mv $THEME_DEST/package.json $THEME_DEST/package.child
-  sed "s/ts_motif/${THEME_NAME}/g" $THEME_DEST/package.child >$THEME_DEST/package.json
-  rm $THEME_DEST/package.child
-
-  # fix up the README file so it reflects the new theme name and location
-  mv $THEME_DEST/README.md $THEME_DEST/README.md.child
-  sed "s/# Motif/# ${THEME_NAME}/g" $THEME_DEST/README.md.child >$THEME_DEST/README.md
-  rm $THEME_DEST/README.md.child
+  for file in $THEME_DEST/*.*
+    do
+      if [[ $file != *.svg ]]
+      then
+        if [[ $file != *.png ]]
+        then
+          echo Updating ${file}
+          mv $file $file.child
+          sed "s/renamed_theme/${THEME_NAME}/g" $file.child > $file
+          rm $file.child
+        fi  
+      fi
+  done
 
   # Rename & edit config files
-  mv $THEME_DEST/config/install/ts_motif.settings.yml $THEME_DEST/config/install/${THEME_NAME}.settings.yml
+  echo Renaming \& updating config/install/renamed_theme.settings.yml
+  mv $THEME_DEST/config/install/renamed_theme.settings.yml $THEME_DEST/config/install/${THEME_NAME}.settings.yml
 
-  for file in $THEME_DEST/config/optional/*ts_motif*yml
+  echo Renaming \& updating configs in config/optional/
+  for file in $THEME_DEST/config/optional/*renamed_theme*yml
     do
-      mv $file $file.child
-      sed "s/ts_motif/${THEME_NAME}/g" $file.child > $file
-      rm $file.child
+      newfile="$(echo ${file} | sed -e "s/renamed_theme/${THEME_NAME}/g")" ;
+      sed "s/renamed_theme/${THEME_NAME}/g" $file > $newfile
+      rm $file
   done
 
   # self modifying code
+  echo Updating this script so it can be run again if you want to change the name again.
   mv $THEME_DEST/update_theme_name.sh $THEME_DEST/update_theme_name.child
-  sed "s/ts_motif/${THEME_NAME}/g" update_theme_name.child | \
-  sed "s/ts_motif/${THEME_NAME}/g" | \
-  sed "s/ts_motif/${THEME_NAME}/g" >$THEME_DEST/update_theme_name.sh
+  sed "s/renamed_theme/${THEME_NAME}/g" update_theme_name.child | \
+  sed "s/renamed_theme/${THEME_NAME}/g" | \
+  sed "s/renamed_theme/${THEME_NAME}/g" >$THEME_DEST/update_theme_name.sh
   chmod a+x $THEME_DEST/update_theme_name.sh
   rm $THEME_DEST/update_theme_name.child
+
+  echo RENAME COMPLETE. Run `cd .`
